Adding exception handler support - fix tests
